Save and Load
Added the BeginGame.cs and SaveAndLoad.cs scripts.  Currently demonstrated in the NewGame scene (which resets the save file) and the TestArea scene, where you can press the up arrow to increase the number of eggs your cube has eaten.  Whenever you load the TestArea Scene, the cube will retain however many eggs it ate, growing to an appropriate size (and starting to move forward for debug purposes).  Basically, just call the IncrementSaveState method whenever the pet egg is fed a food egg, and we'll be able to alternate between scenes freely.