A Bunch of Stuff
I added an egg to the tree, and I added animations, an animator, and code to make the egg fall down when space is pressed.  You now type "heck" instead of "h" to access the minigame.  Added instructional paintings of "space" and "t".  Since the egg tree egg was accidentally adding 2 eggs every time, I jsut doubled the egg scale.